backend/requirements.txt
fastapi==0.116.0
uvicorn[standard]==0.35.0
python-dotenv==1.1.1
pydantic==2.11.7
sse-starlette==2.3.6
langchain==1.2.15
langchain-openai==1.1.12
httpx==0.28.1
pytest==8.4.1
pytest-asyncio==1.1.0
rank-bm25==0.2.2
